+++ src/features/js_fetch/proxies.rs
@@ -1,9 +1,9 @@
 use crate::features::js_fetch::spec::Response;
-use hirofa_utils::auto_id_map::AutoIdMap;
 use hirofa_utils::js_utils::adapters::proxies::JsProxy;
 use hirofa_utils::js_utils::adapters::JsRealmAdapter;
 use hirofa_utils::js_utils::JsError;
 use std::cell::RefCell;
+use std::collections::HashMap;
 use std::sync::Arc;
 
 pub(crate) fn impl_for<C>(realm: &C) -> Result<(), JsError>
@@ -15,7 +15,7 @@ where
 }
 
 thread_local! {
-    static RESPONSE_INSTANCES: RefCell<AutoIdMap<Arc<Response>>> = RefCell::new(AutoIdMap::new());
+    pub(crate) static RESPONSE_INSTANCES: RefCell<HashMap<usize, Arc<Response>>> = RefCell::new(HashMap::new());
 }
 
 fn with_response<C: FnOnce(&Arc<Response>) -> R, R>(id: &usize, consumer: C) -> Result<R, &str> {
@@ -34,7 +34,8 @@ where
     C: JsRealmAdapter + 'static,
 {
     let mut response_proxy = JsProxy::new(&[], "Response");
-    response_proxy.set_finalizer(|rt, realm, id| {
+    response_proxy.set_finalizer(|_rt, _realm, id| {
+        // todo.. need to use realm id as part of key?
         RESPONSE_INSTANCES.with(|rc| {
             let map = &mut *rc.borrow_mut();
             map.remove(id);
@@ -46,12 +47,20 @@ where
             .map_err(|s| JsError::new_str(s))?;
         // todo promise may seem futile now but later we will just store bytes in body and encode to string async
         realm.js_promise_create_resolving(
-            async move {
-                let txt = response.body.text.clone(); // todo impl take in body
-                Ok(txt)
-            },
+            async move { response.text().await },
             // todo js_string_crea2 with String
             |realm, res| realm.js_string_create(res.as_str()),
+        )
+    });
+    response_proxy.add_method("json", |_rt, realm: &C, instance_id, _args| {
+        //
+        let response = with_response(instance_id, |response| response.clone())
+            .map_err(|s| JsError::new_str(s))?;
+        // todo promise may seem futile now but later we will just store bytes in body and encode to string async
+        realm.js_promise_create_resolving(
+            async move { response.text().await },
+            // todo js_string_crea2 with String
+            |realm, res| realm.js_json_parse(res.as_str()),
         )
     });