+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * ifdef preprocessor, uses [gpp](https://github.com/Kestrer/gpp)
 * added abstract console
+* removed old fetchresponseprovider and added new Fetch api (wip)
  
 0.0.1